requirements.txt
django-user-accounts==1.3.1
pinax-webanalytics==2.0.4
pinax-eventlog==1.1.2

Django==1.11.3
pinax-theme-bootstrap==7.10.1
pillow==4.2.1
pandas==0.20.3
bokeh==0.12.9
scipy==0.19.1
pylatex=1.2.1
selenium=3.6.0
